DDA-83: Refactor graph filter
